add App component after login
